fix: calculate win position according to width and height
If we calculate the window position according to the default width and
height, it's always going to be fixed at a single point in the editor
grid. If the user changes the width and height, then the window is not
guarranteed to be in the center.

This will fix it by calculating the position according to the given
width and height, be it the default one or the user provided one.